Plasmid sequences now removed from target genomes.

+++ Core.py
@@ -1,13 +1,11 @@
 from argparse import ArgumentParser
-from collections import defaultdict
 import glob
 from USSPpip import SigSeekr
 import os
 import shutil
-import json
 import gzip
-import GeneSeekrUper as GeneSeekr
 from All2AllMash import run_mash, read_mash
+import subprocess
 
 
 __author__ = 'mikeknowles'
@@ -18,46 +16,6 @@ To sort rMLST results and remove closely related sequences
 then to prepare the data for strain-specific probe idenification
 """
 
-def retriever(genomes, output):
-    if not os.path.exists(output + "Genomes"):
-        os.mkdir(output + "Genomes")
-    for folders in glob.glob(genomes + "/*"):
-        if os.path.exists(folders + "/Best_Assemblies"):
-            for fasta in glob.glob(folders + "/Best_Assemblies/*"):
-                shutil.copy(fasta, output + "Genomes")
-
-def jsonUpGoer(jsonfile, markers, genomes, outdir, method):
-    if os.path.isfile(jsonfile):
-        genedict = json.load(open(jsonfile))
-    else:
-        genedict = GeneSeekr.blaster(outdir, 100, genomes, markers, "USSpip_" + method)
-        # genedict = GeneSeekr.blaster(markers, genomes, outdir, "USSpip_" + method)
-        handle = open(jsonfile, 'w')
-        json.dump(genedict, handle, sort_keys=True, indent=4, separators=(',', ': '))
-        handle.close()
-    return genedict
-
-def alleledictbuild(TargetrMLST, nonTargetrMLST):
-    typing = defaultdict(dict)
-    for genome in TargetrMLST:  # genome refers to target genome
-        if genome not in typing:  # add the target genome to the dictionary
-            typing[genome] = defaultdict(int)
-
-        for gene in sorted(TargetrMLST[genome]):  # gene is the rST
-            for nontarget in nonTargetrMLST:  # check against this genome
-                if nontarget not in typing[genome]:  # if nontarget genome not in typing dictionary then add it
-                    typing[genome][nontarget] = 0
-                if gene in TargetrMLST[genome] and gene in nonTargetrMLST[nontarget]:  #
-                    match = 0
-                    for allele in TargetrMLST[genome][gene]:  # multiple allele types possibly present
-                        if allele in nonTargetrMLST[nontarget][gene]:
-                            match += 1
-                    if match == len(nonTargetrMLST[nontarget][gene]):
-                        typing[genome][nontarget] += 1
-    return typing
-
-    # return typing, removed
-
 
 def uncompress_file(filename):
     in_gz = gzip.open(filename, 'rb')
@@ -65,6 +23,13 @@ def uncompress_file(filename):
     out.write(in_gz.read())
     out.close()
 
+
+def remove_plasmid_sequences(target_file):
+    cmd = 'bbduk.sh ref=plasmid_database.fa in={} out={} overwrite'.format(target_file, target_file.replace('.fasta', '_noplasmids.fasta'))
+    print('Removing plasmid sequences...')
+    print(cmd)
+    subprocess.call(cmd, shell=True)
+    return target_file.replace('.fasta', '_noplasmids.fasta')
 
 def sorter(genomes, outdir, target, evalue, estop, mash_cutoff, threads):
     # if not os.path.exists(outdir + "Genomes/"):
@@ -77,29 +42,18 @@ def sorter(genomes, outdir, target, evalue, estop, mash_cutoff, threads):
     # nontargets = glob(genomes + "*.fa*")
     to_remove = list()
     nontargets = glob.glob(genomes + '/*.f*a*')
-    for i in range(len(nontargets)):
-        if nontargets[i].endswith('.gz'):
-            uncompress_file(nontargets[i])
-            to_remove.append(nontargets[i].replace('.gz', ''))
-            nontargets[i] = nontargets[i].replace('.gz', '')
-    run_mash(genomes, threads)
-    nontargets = read_mash("tmp/distances.txt", mash_cutoff)
-    shutil.rmtree("tmp/")
-    # jsonfile = '%sgenedict.json' % outdir
-    # nonTargetrMLST = jsonUpGoer(jsonfile, markers, genomes, outdir, 'nontarget')
-    # if os.path.exists(target):  # Determine if target is a folder
-    #    targetjson = '%stargetdict.json' % outdir
-    #else:
-    #    print "The variable \"--targets\" is not a folder or file "
-    #    return
-    # TargetrMLST = jsonUpGoer(targetjson, markers, target, outdir, 'target')
-    # typing, removed = compareType(TargetrMLST, nonTargetrMLST)
-    # json.dump(typing, open(outdir + 'typing.json', 'w'), sort_keys=False, indent=4, separators=(',', ': '))
-    # json.dump(removed, open(outdir + 'removed.json', 'w'), sort_keys=False, indent=4, separators=(',', ': '))
-    # for sigtarget in typing:
-    #    print typing
-    #    print typing[sigtarget]
-    #    SigSeekr(typing, typing[sigtarget], outdir, evalue, float(estop), 200, 1)
+    if len(nontargets) > 1:
+        for i in range(len(nontargets)):
+            if nontargets[i].endswith('.gz'):
+                uncompress_file(nontargets[i])
+                to_remove.append(nontargets[i].replace('.gz', ''))
+                nontargets[i] = nontargets[i].replace('.gz', '')
+        run_mash(genomes, threads)
+        nontargets = read_mash("tmp/distances.txt", mash_cutoff)
+        shutil.rmtree("tmp/")
+    # If target is a folder, try to find something matching to any of the files by using a concatenated
+    # sequence...not exactly what I meant to do with this. Probably need to actually go through and find a core
+    # genome to get a sequence that's common to all of the files as you previously intended.
     if os.path.isdir(target):
         fasta_files = glob.glob(target + '/*.f*a*')
         for fasta in fasta_files:
@@ -112,12 +66,18 @@ def sorter(genomes, outdir, target, evalue, estop, mash_cutoff, threads):
             # SigSeekr(fasta, nontargets, outdir, float(evalue), float(estop), 200, 1)
         target = outdir + '/concatenated_target.fasta'
         to_remove.append(target)
+        target = remove_plasmid_sequences(target)
+        to_remove.append(target)
     else:
         if '.gz' in target:
             uncompress_file(target)
             to_remove.append(target.replace('.gz', ''))
             target = target.replace('.gz', '')
-
+            target = remove_plasmid_sequences(target)
+            to_remove.append(target)
+        else:
+            target = remove_plasmid_sequences(target)
+            to_remove.append(target)
     # else:
     SigSeekr(target, nontargets, outdir, float(evalue), float(estop), 200, 1)
     for item in to_remove:
